requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-dotenv==1.0.0
httpx==0.24.1
pydantic==2.0.3
pydantic-settings==2.0.3
python-multipart==0.0.6

# 测试依赖
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
requests==2.31.0 